% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CauMod.R
\name{CauMod}
\alias{CauMod}
\title{Cauchy autocorrelation model}
\usage{
CauMod(dist, rang)
}
\arguments{
\item{dist}{a vector or matrix of distance values}

\item{rang}{the range of the spherical model}
}
\value{
a vector or matrix of the Cauchy model values for each distance
}
\description{
Cauchy autocorrelation model
}
\author{
Jay Ver Hoef
}
