% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossVal.R
\name{crossVal}
\alias{crossVal}
\title{Crossvalidation from a fitted spatial model}
\usage{
crossVal(z, X, V, Vi, n, p, covb, covbi, bhat)
}
\arguments{
\item{z}{vector of data}

\item{X}{design matrix}

\item{V}{covariance matrix}

\item{Vi}{inverse of covariance matrix}

\item{n}{number of observations, number of rows in z}

\item{p}{number of covariates, number of columns in X}

\item{covb}{covariance matrix of fitted fixed effects}

\item{covbi}{inverse of covb}

\item{bhat}{fitted fixed effects}
}
\value{
a data.frame with crossvalidation predictions "cv.pred" and estimated prediction standard errors "cv.se"
}
\description{
Crossvalidation from a fitted spatial model
}
\author{
Jay Ver Hoef
}
