% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CVsummStats.R
\name{CVsummStats}
\alias{CVsummStats}
\title{Summary statistics from a leave-one-out crossvalidation (LOOCV) from a 
fitted spatial model}
\usage{
CVsummStats(optimFit, obsVec, CVtable)
}
\arguments{
\item{optimFit}{the optimized likelihood using the optim() function}

\item{CVtable}{an n x 2 table from LOOCV, where the predicted values are in
the first column, and the prediction variances are in the second column.}

\item{obsvec}{a vector of the observed data}
}
\value{
a list with [[1]] AIC; [[2]] corr: the correlation between the obseved 
values and the LOOCV predictions; [[3]] RMSPE, the root-mean-squared prediction 
error from LOOCV; and [[4]] CI90, the proportion of times the LOOCV 90/%
prediction interval covers the true value.
}
\description{
Summary statistics from a leave-one-out crossvalidation (LOOCV) from a 
fitted spatial model
}
\author{
Jay Ver Hoef
}
