% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empsvgm.R
\name{empsvgm}
\alias{empsvgm}
\title{Empirical semivariogram}
\usage{
empsvgm(z, distMat, breaks)
}
\arguments{
\item{z}{a vector of the observed data}

\item{distMat}{distance matrix between all pairs of locations in the z vector,
in the same order as the z vector}

\item{breaks}{a vector of breaks to create distance bins}
}
\value{
a data.frame with 3 columes.  The first column is empirical
semivariogram for each distance class, the second column is the mean
distance in that distance class, and the final column is the number
of pairs of points in that distance class
}
\description{
Empirical semivariogram
}
\author{
Jay Ver Hoef
}
