% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExpMod.R
\name{ExpMod}
\alias{ExpMod}
\title{Exponential autocorrelation model}
\usage{
ExpMod(dist, rang)
}
\arguments{
\item{dist}{a vector or matrix of distance values}

\item{rang}{the range of the exponential model}
}
\value{
a vector or matrix of the exponential model values for each distance
}
\description{
Exponential autocorrelation model
}
\author{
Jay Ver Hoef
}
