% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GauMod.R
\name{GauMod}
\alias{GauMod}
\title{Gaussian autocorrelation model}
\usage{
GauMod(dist, rang)
}
\arguments{
\item{dist}{a vector or matrix of distance values}

\item{rang}{the range of the Gaussian model}
}
\value{
a vector or matrix of the Gaussian model values for each distance
}
\description{
Gaussian autocorrelation model
}
\author{
Jay Ver Hoef
}
