% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HolMod.R
\name{HolMod}
\alias{HolMod}
\title{Hole effect autocorrelation model}
\usage{
HolMod(dist, rang)
}
\arguments{
\item{dist}{a vector or matrix of distance values}

\item{rang}{the range of the hole effect model}
}
\value{
a vector or matrix the hole effect model values for each distance
}
\description{
Hole effect autocorrelation model
}
\author{
Jay Ver Hoef
}
