% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LOOCV.R
\name{LOOCV}
\alias{LOOCV}
\title{Leave-one-out crossvalidation (LOOCV) for a fitted spatial model}
\usage{
LOOCV(SigMat, obsVec)
}
\arguments{
\item{SigMat}{The fitted covariance matrix for the spatial data}

\item{obsvec}{a vector of the observed data}
}
\value{
a data.frame with the predictions in the first column and the 
prediction variances in the second column
}
\description{
Leave-one-out crossvalidation (LOOCV) for a fitted spatial model
}
\author{
Jay Ver Hoef
}
