% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LL.R
\name{m2LL}
\alias{m2LL}
\title{minus 2 times log-likelihood for spatial covariance model}
\usage{
m2LL(theta, z, Dist, aCorMod, estMeth = "REML")
}
\arguments{
\item{theta}{vector of estimated covariance parameters, where theta[1] is the partial sill, theta[2] is the range, 
and theta[3] is the nugget effect}

\item{z}{vector of data}

\item{aCorMod}{autocorrelation model}

\item{estMeth}{estimation method.  Default is "REML" for restricted maximum likelihood.  Other options are "ML" 
for maximum likelihood}

\item{EucDist}{matrix of Euclidean distances}
}
\value{
minus 2 times the loglikelihood
}
\description{
minus 2 times log-likelihood for spatial covariance model
}
\author{
Jay Ver Hoef
}
