% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LLexp.R
\name{m2LLexp}
\alias{m2LLexp}
\title{minus 2 times log-likelihood for exponential spatial covariance model}
\usage{
m2LLexp(theta, z, EucDist, estMeth = "REML")
}
\arguments{
\item{theta}{vector of estimated covariance parameters, where theta[1] is the partial sill, theta[2] is the range, 
and theta[3] is the nugget effect}

\item{z}{vector of data}

\item{EucDist}{matrix of Euclidean distances}

\item{estMeth}{estimation method.  Default is "REML" for restricted maximum likelihood.  Other options are "ML" 
for maximum likelihood}
}
\value{
minus 2 times the loglikelihood
}
\description{
minus 2 times log-likelihood for exponential spatial covariance model
}
\author{
Jay Ver Hoef
}
