% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LLrr.R
\name{m2LLrr}
\alias{m2LLrr}
\title{minus 2 times log-likelihood for reduced rank models using Sherman-Morrison-Woodbury}
\usage{
m2LLrr(theta, z, rrDist, knDist, corMod = "exp", estMeth = "REML")
}
\arguments{
\item{theta}{vector of estimated covariance parameters, with theta[1] the partial sill, 
theta[2] the range parameter of the reduced rank matrix, theta[3] the range parameter for the knots, and 
theta[4] the nugget effect.}

\item{z}{vector of data}

\item{rrDist}{distance matrix with any kind of distance metric. The reduced rank distance matrix.}

\item{knDist}{Euclidean distance matrix among knot location}

\item{corMod}{one of the autocorrelation models.  Valid options are 'exp', 'sph', 'gau', and 'cau', for 
exponential, spherical, Gaussian, and Cauchy, respectively}

\item{estMeth}{estimation method.  Default is "REML" for restricted maximum likelihood.  Other options are "ML" for maximum likelihood}
}
\value{
minus 2 times the (restricted) loglikelihood
}
\description{
minus 2 times loglikelihood for reduced rank models using Sherman-Morrison-Woodbury
}
\author{
Jay Ver Hoef
}
