% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/m2LLvcSph.R
\name{m2LLvcSph}
\alias{m2LLvcSph}
\title{minus 2 times log-likelihood for variance component model}
\usage{
m2LLvcSph(theta, z, rrDist, knDist, EdisMat, estMeth = "REML")
}
\arguments{
\item{theta}{vector of estimated covariance parameters, where theta[1] is the 
partial sill for linear distance, theta[2] is the range for linear distance,
theta[3] is the partial sill Euclidean distance, theta[4] is the range for 
knots when using linear distance, theta[5] is the nugget effect, and theta[6]
is the range for Euclidean distance}

\item{z}{vector of data}

\item{rrDist}{linear network distance matrix between knots and observed data}

\item{knDist}{Euclidean distance matrix among knots}

\item{EdisMat}{matrix of Euclidean distances}

\item{estMeth}{estimation method.  Default is "REML" for restricted maximum 
likelihood.  Other options are "ML" for maximum likelihood}
}
\value{
minus 2 times the loglikelihood
}
\description{
minus 2 times log-likelihood for variance component model with two spherical 
models, one using the reduced rank model for linear distance, and the other 
using Euclidean distance
}
\author{
Jay Ver Hoef
}
