% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makeFigNetwork127.R
\name{makeFigNetwork127}
\alias{makeFigNetwork127}
\title{makes stream network figure, part (c) in Figure 2}
\usage{
makeFigNetwork127()
}
\value{
a plot
}
\description{
makes stream network figure, part (c) in Figure 2
}
\author{
Jay Ver Hoef
}
