% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/minEigVal.R
\name{minEigVal}
\alias{minEigVal}
\title{minimum eigenvalue from a spatial covariance matrix}
\usage{
minEigVal(fun.acor, Dis, alpha)
}
\arguments{
\item{Dis}{distance matrix}

\item{alpha}{range parameter of autocorrelation function}

\item{fun.cor}{one of the autocorrelation functions}
}
\value{
minimum eigenvalue
}
\description{
minimum eigenvalue from a spatial covariance matrix
}
\author{
Jay Ver Hoef
}
