% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SMWXy.R
\name{SMWXy}
\alias{SMWXy}
\title{Sherman Morrison Woodbury inverse}
\usage{
SMWXy(Ainv, Alogdet, BinvZ, Blogdet, Z, BinvX, Binvy)
}
\arguments{
\item{Ainv}{inverse of A matrix}

\item{Alogdet}{determinant of A matrix}

\item{BinvZ}{inverse of B matrix times Z}

\item{Blogdet}{log determinant of B matrix}

\item{Z}{a matrix}

\item{BinvX}{inverse of B matrix times X}

\item{Binvy}{inverse of B matrix times y}
}
\value{
a list with the inverse inv(Z A t(Z) + B) X, inv(Z A t(Z) + B) y and determinant of Z A t(Z) + B
}
\description{
Uses Sherman Morrison Woodbury on matrix formula inv(Z A t(Z) + B) and returns inv(Z A t(Z) + B) X and inv(Z A t(Z) + B) y  
and the log determinant. This keeps from storing the whole inverse for very large data problems.
}
\author{
Jay Ver Hoef \email{jay.verhoef@noaa.gov}
}
