% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SphMod.R
\name{SphMod}
\alias{SphMod}
\title{Spherical autocorrelation model}
\usage{
SphMod(dist, rang)
}
\arguments{
\item{dist}{a vector or matrix of distance values}

\item{rang}{the range of the spherical model}
}
\value{
a vector or matrix of the spherical model values for each distance
}
\description{
spherical autocorrelation model
}
\author{
Jay Ver Hoef
}
