% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vgrmFitWLS.R
\name{vgrmFitWLS}
\alias{vgrmFitWLS}
\title{Weighted least squares function for a variogram model}
\usage{
vgrmFitWLS(theta, aCorMod, empsvgm, dist, npair, useNugget = TRUE)
}
\arguments{
\item{theta}{a vector of covariance parameters, with theta[1] the partial 
sill, theta[2] the range parameter, and theta[3] the nugget effect, on 
the log scale}

\item{aCorMod}{an autocorrelation model that takes a distance matrix and
returns a matrix of autocorrelation values.  The autocorrelation model
should have 2 arguments, with the first the distance matrix, the second
the range parameter.}

\item{empsvgm}{a vector of empirical semivariogram values}

\item{dist}{a vector of distances matched to (the same length as) the 
semivariogram values}

\item{npair}{a vector of the number of location-pairs matched to (the same 
length as) the semivariogram values}

\item{useNugget}{Either TRUE or FALSE on whether to add a nugget effect
to the model.  If TRUE, the theta argument must have a length of 3.
Default is TRUE.}
}
\value{
the weighted sums-of-squares of the difference between the empirical
  semivariogram values and the fitted model for the given theta.
}
\description{
Weighted least squares function for a variogram model
}
\author{
Jay Ver Hoef
}
